feat(favorites): интегрировать FavoritesViewModel и переработать архитектуру избранного
**Рефакторинг структуры базы данных**
- Перемещены классы базы данных в правильный пакет core.data.database
  - AppDatabase, FavoritesDao, FavoriteVacancyEntity теперь в core
- Исправлено название метода DAO: favoriteDao() → favoritesDao()
- Обновлены импорты во всех зависимых файлах

**Модуль DI**
- В PresentationModule добавлена регистрация FavoritesViewModel
- DataModule использует обновленные пути к классам базы данных
- FavoritesRepository теперь импортируется из core.domain.repository

**FavoritesViewModel**
- Реализован ViewModel, который подписывается на Flow из репозитория
- Преобразует список вакансий в FavoritesUiState (Content/Empty)
- Использует SharingStarted.WhileSubscribed с таймаутом 5 секунд
- Начальное состояние — Loading

**Навигация**
- Удалена временная тестовая логика из NavigationRoot
- FavoritesScreen теперь получает state из ViewModel через collectAsState()
- Обработчик onVacancyClick теперь просто добавляет маршрут Vacancy в бэкстек
- Убран прямой вызов репозитория и корутин

**FavoritesScreen**
- Исправлен импорт FavoritesUiState (теперь из пакета presentation)
