+++ app/src/main/java/ru/practicum/android/diploma/app/di/PresentationModule.kt
@@ -3,6 +3,8 @@ package ru.practicum.android.diploma.app.di
 import org.koin.core.module.dsl.viewModel
 import org.koin.dsl.module
 import ru.practicum.android.diploma.app.navigation.NavigationViewModel
+import ru.practicum.android.diploma.core.domain.repository.FavoritesRepository
+import ru.practicum.android.diploma.feature.favorite.presentation.FavoritesViewModel
 
 /**
  * Модуль Koin, отвечающий за зависимости UI и ViewModel
@@ -11,6 +13,12 @@ val presentationModule = module {
 
     viewModel<NavigationViewModel> {
         NavigationViewModel()
+    }
+
+    viewModel<FavoritesViewModel> {
+        FavoritesViewModel(
+            get<FavoritesRepository>()
+        )
     }
 
 }
