+++ app/src/main/java/ru/practicum/android/diploma/core/data/database/db/AppDatabase.kt
@@ -2,14 +2,16 @@ package ru.practicum.android.diploma.core.data.database.db
 
 import androidx.room.Database
 import androidx.room.RoomDatabase
+import androidx.room.TypeConverters
 import ru.practicum.android.diploma.core.data.database.dao.FavoritesDao
 import ru.practicum.android.diploma.core.data.database.entity.FavoriteVacancyEntity
+import ru.practicum.android.diploma.feature.favorite.data.StringListTypeConverter
 
 @Database(
-    entities = [FavoriteVacancyEntity::class],
-    version = 1
+    version = 1,
+    entities = [FavoriteVacancyEntity::class]
 )
+@TypeConverters(StringListTypeConverter::class)
 abstract class AppDatabase : RoomDatabase() {
-
     abstract fun favoritesDao(): FavoritesDao
 }
