+++ app/src/main/java/ru/practicum/android/diploma/core/data/database/entity/FavoriteVacancyEntity.kt
@@ -1,18 +1,28 @@
 package ru.practicum.android.diploma.core.data.database.entity
 
+import androidx.room.Embedded
 import androidx.room.Entity
 import androidx.room.PrimaryKey
+import ru.practicum.android.diploma.core.config.DatabaseConfig
+import ru.practicum.android.diploma.core.domain.model.Address
+import ru.practicum.android.diploma.core.domain.model.Contacts
+import ru.practicum.android.diploma.core.domain.model.Employer
+import ru.practicum.android.diploma.core.domain.model.Salary
 
-@Entity(tableName = "favorites")
+@Entity(tableName = DatabaseConfig.FAVORITE_VACANCY_TABLE)
 data class FavoriteVacancyEntity(
-
-    @PrimaryKey
-    val id: String,
+    @PrimaryKey val id: String,
     val name: String,
-    val employer: String,
-    val salaryFrom: Int?,
-    val salaryTo: Int?,
-    val currency: String?,
-    val city: String,
-    val url: String
+    val description: String,
+    @Embedded(prefix = "salary_") val salary: Salary?,
+    @Embedded(prefix = "address_") val address: Address?,
+    val experience: String?,
+    val schedule: String?,
+    val employmentType: String?,
+    @Embedded(prefix = "contacts_") val contacts: Contacts?,
+    @Embedded(prefix = "employer_") val employer: Employer,
+    val skills: List<String>,
+    val website: String,
+    val industry: String,
+    val addedAtMillis: Long = System.currentTimeMillis()
 )
