+++ app/src/main/java/ru/practicum/android/diploma/core/domain/repository/FavoritesRepository.kt
@@ -1,12 +1,9 @@
 package ru.practicum.android.diploma.core.domain.repository
 
+import kotlinx.coroutines.flow.Flow
 import ru.practicum.android.diploma.core.domain.model.Vacancy
+import ru.practicum.android.diploma.feature.vacancy.data.FavoritesDataSource
 
-interface FavoritesRepository {
-
-    suspend fun addFavorite(vacancy: Vacancy)
-
-    suspend fun removeFavorite(id: String)
-
-    suspend fun getFavorites(): List<Vacancy>
+interface FavoritesRepository : FavoritesDataSource {
+    fun getVacancies(): Flow<List<Vacancy>>
 }
