--- app/src/main/java/ru/practicum/android/diploma/feature/favorite/domain/FavoritesRepository.kt
@@ -1,9 +1,0 @@
-package ru.practicum.android.diploma.feature.favorite.domain
-
-import kotlinx.coroutines.flow.Flow
-import ru.practicum.android.diploma.core.domain.model.Vacancy
-import ru.practicum.android.diploma.feature.vacancy.data.FavoritesDataSource
-
-interface FavoritesRepository : FavoritesDataSource {
-    fun getVacancies(): Flow<List<Vacancy>>
-}
