+++ app/src/main/java/ru/practicum/android/diploma/feature/favorite/presentation/FavoritesViewModel.kt
@@ -1,0 +1,33 @@
+package ru.practicum.android.diploma.feature.favorite.presentation
+
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.viewModelScope
+import kotlinx.coroutines.flow.SharingStarted
+import kotlinx.coroutines.flow.StateFlow
+import kotlinx.coroutines.flow.map
+import kotlinx.coroutines.flow.stateIn
+import ru.practicum.android.diploma.core.domain.repository.FavoritesRepository
+
+class FavoritesViewModel(
+    repository: FavoritesRepository
+) : ViewModel() {
+
+    companion object {
+        private const val TIMEOUT_MILLIS = 5000L
+    }
+
+    val state: StateFlow<FavoritesUiState> =
+        repository.getVacancies()
+            .map { vacancies ->
+                when {
+                    vacancies.isEmpty() -> FavoritesUiState.Empty
+                    else -> FavoritesUiState.Content(vacancies)
+                }
+            }
+            .stateIn(
+                scope = viewModelScope,
+                started = SharingStarted.WhileSubscribed(TIMEOUT_MILLIS),
+                initialValue = FavoritesUiState.Loading
+            )
+
+}
